+++ flatlaf-testing/src/main/java/com/formdev/flatlaf/testing/FlatTestFrame.java
@@ -455,8 +455,12 @@ public class FlatTestFrame
 	}
 
 	private void inspectChanged() {
-		if( inspector == null )
-			inspector = new FlatInspector( contentPanel );
+		if( inspector == null ) {
+			inspector = new FlatInspector( getRootPane() );
+			inspector.addPropertyChangeListener( e -> {
+				inspectCheckBox.setSelected( inspector.isEnabled() );
+			} );
+		}
 		inspector.setEnabled( inspectCheckBox.isSelected() );
 	}